+++ angel-ps/core/src/main/java/com/tencent/angel/ml/math2/storage/IntIntSortedVectorStorage.java
@@ -35,7 +35,7 @@ public class IntIntSortedVectorStorage implements IntIntVectorStorage {
   }
 
   public IntIntSortedVectorStorage(int dim) {
-    this(dim, Math.min(64, (int) dim));
+    this(dim, Math.min(64, dim));
   }
 
   @Override public int get(int idx) {
