+++ angel-ps/core/src/main/java/com/tencent/angel/ml/math2/storage/IntLongSortedVectorStorage.java
@@ -35,7 +35,7 @@ public class IntLongSortedVectorStorage implements IntLongVectorStorage {
   }
 
   public IntLongSortedVectorStorage(int dim) {
-    this(dim, Math.min(64, (int) dim));
+    this(dim, Math.min(64, dim));
   }
 
   @Override public long get(int idx) {
